requirements.txt
datasets>=2.14.2
rouge-score>=0.0.4
nlpaug>=1.1.10
scikit-learn
tqdm>=4.64.1
matplotlib>=3.6
pandas>=1.3.5
torch>=1.13.0
bs4
transformers<=4.36
nltk>=3.6.5
sacrebleu>=1.5.0
sentencepiece>=0.1.97
hf-lfs==0.0.3
pytest>=4.4.1
pytreebank>=0.2.7
setuptools>=60.2.0
numpy>=1.23.5
dill>=0.3.5.1
scipy>=1.9.3
flask>=2.3.2
protobuf>=4.23
fschat>=0.2.3
hydra-core>=1.3.2
einops
accelerate
bitsandbytes
openai==0.28.0
wget
sentence-transformers
bert-score
